fix: resolve auth api and dashboard issues
