+++ frontend/.env.example
@@ -1,3 +1,6 @@
-AUTH_SECRET={AUTH_SECRET}
+AUTH_SECRET=XKZ7Z1O3bhU4caz2D8ouu7dKhXyQ+ZUhsjvTwtHxp10=
+NEXTAUTH_SECRET=XKZ7Z1O3bhU4caz2D8ouu7dKhXyQ+ZUhsjvTwtHxp10=
+AUTH_URL=http://localhost:3001
+NEXTAUTH_URL=http://localhost:3001
 AUTH_GOOGLE_ID={AUTH_GOOGLE_ID}
 AUTH_GOOGLE_SECRET={AUTH_GOOGLE_SECRET}
